refactor/#85: parent/schedule 경로 변경
