+++ app/src/main/java/com/kiero/presentation/parent/mission/component/datepicker/component/CalendarMontHeader.kt
@@ -1,4 +1,4 @@
-package com.kiero.presentation.parent.schedule.mission.component.datepicker.component
+package com.kiero.presentation.parent.mission.component.datepicker.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
