+++ app/src/main/java/com/kiero/presentation/parent/mission/component/datepicker/component/DaysOfWeekTitle.kt
@@ -1,4 +1,4 @@
-package com.kiero.presentation.parent.schedule.mission.component.datepicker.component
+package com.kiero.presentation.parent.mission.component.datepicker.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -14,7 +14,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.kiero.core.designsystem.theme.KieroTheme
-import com.kiero.presentation.parent.schedule.mission.component.datepicker.util.daysOfWeek
+import com.kiero.presentation.parent.mission.component.datepicker.util.daysOfWeek
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
 import java.time.DayOfWeek
