+++ app/src/main/java/com/kiero/presentation/parent/mission/component/datepicker/model/CalendarDay.kt
@@ -1,4 +1,4 @@
-package com.kiero.presentation.parent.schedule.mission.component.datepicker.model
+package com.kiero.presentation.parent.mission.component.datepicker.model
 
 import androidx.compose.runtime.Immutable
 import java.time.LocalDate
