--- app/src/main/java/com/kiero/presentation/parent/schedule/mission/component/datepicker/model/DownTimeStatus.kt
@@ -1,8 +1,0 @@
-package com.kiero.presentation.parent.schedule.mission.component.datepicker.model
-
-
-enum class DateState {
-    PAST,
-    TODAY,
-    FUTURE
-}
